+++ packages/cqrs/test/cqrs_test.dart
@@ -155,6 +155,81 @@ void main() {
         );
       });
     });
+
+    group('perform', () {
+      test(
+          "correctly serializes operation, calls client's send and"
+          ' deserializes result', () async {
+        mockClientPost(client, Response('true', 200));
+
+        final result = await cqrs.perform(
+          ExampleOperation(),
+          headers: {'X-Test': 'foobar'},
+        );
+
+        expect(result, true);
+
+        verify(
+          client.post(
+            Uri.parse('https://example.org/api/operation/ExampleOperation'),
+            body: anyNamed('body'),
+            headers: argThat(
+              isA<Map<String, String>>()
+                  .having((h) => h['X-Test'], 'X-Test', 'foobar')
+                  .having(
+                    (h) => h['Content-Type'],
+                    'Content-Type',
+                    'application/json',
+                  ),
+              named: 'headers',
+            ),
+          ),
+        ).called(1);
+      });
+
+      test('correctly deserializes null operation result', () async {
+        mockClientPost(client, Response('null', 200));
+
+        final result = await cqrs.perform(ExampleOperation());
+
+        expect(result, null);
+      });
+
+      test('throws CQRSException on json decoding failure', () async {
+        mockClientPost(client, Response('true', 200));
+
+        final result = cqrs.perform(ExampleOperationFailingResultFactory());
+
+        expect(
+          result,
+          throwsA(
+            isA<CqrsException>().having(
+              (e) => e.message,
+              'message',
+              'An error occured while decoding response body JSON:\n'
+                  'Exception: This is error.',
+            ),
+          ),
+        );
+      });
+
+      test('throws CQRSException when response code is other than 200', () {
+        mockClientPost(client, Response('', 404));
+
+        final result = cqrs.perform(ExampleOperation());
+
+        expect(
+          result,
+          throwsA(
+            isA<CqrsException>().having(
+              (e) => e.message,
+              'message',
+              'Invalid, non 200 status code returned by ExampleOperation operation.',
+            ),
+          ),
+        );
+      });
+    });
   });
 }
 
@@ -182,6 +257,28 @@ class ExampleQuery extends Query<bool?> {
 class ExampleQueryFailingResultFactory extends Query<bool> {
   @override
   String getFullName() => 'ExampleQuery';
+
+  @override
+  bool resultFactory(dynamic json) => throw Exception('This is error.');
+
+  @override
+  Map<String, dynamic> toJson() => <String, dynamic>{};
+}
+
+class ExampleOperation extends Operation<bool?> {
+  @override
+  String getFullName() => 'ExampleOperation';
+
+  @override
+  bool? resultFactory(dynamic json) => json as bool?;
+
+  @override
+  Map<String, dynamic> toJson() => <String, dynamic>{};
+}
+
+class ExampleOperationFailingResultFactory extends Operation<bool> {
+  @override
+  String getFullName() => 'ExampleOperation';
 
   @override
   bool resultFactory(dynamic json) => throw Exception('This is error.');